Fix #1
We're now using events to manage the syncing state.

We could probably do the same without the events but I wanted to make sure the messages when starting/stopping would be logged immediately.

There was a bug where the "auto syncing stopped" message would take a second to appear when clicking the Sync button to toggle the plugin off. This fixes that, as well.